requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
djangorestframework==3.15.2
idna==3.8
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
